added configuration elements as preferences
set offsets via preferences, not sure if working but they are being recorded and remembered and processed now without error. this is needed as RH is low by 7% and temp is high by 6 degrees F. not sure if UV sensor is working or not. Will query Aeotech, went by sensor multilevel v5 documentation for type 27 ultraviolet, which may not be accurate in their mapping. haven't seen any activity on it but it does pull a zero value.