add acceleration capability
provides vibration sensor output to smartthings in the form of the acceleration capability for use in the home security section. You can create push notifications, SMS, and trigger cameras and the like if you sensor is tampered with or it detects a significant enough vibration, like a door slamming close by.