v1.2 push
corrected the terminology in preferences. what were previously described
as offsets are actually thresholds for reporting. if you set temp to 50
for example, it will report if temp changes at least a half a degree.
humidity is in 1 percent increments, light as well.. also changed wake
up interval to 60 minutes at which point the battery level will also be
reported and if any configuration changes have been requested, will also
be sent to the device. This should fix the lack of battery reporting.