new device type for GE/Jasco Z-Wave switches
adds preferences for switch inversion. I had a switch that randomly
inverted itself (down was on, up was off) and needed a way to set this
manufacturer option, so I created this devicetype off the stock
SmartThings one. if you need to control inversion, edit the device
preferences from the app and set it to true to make up be off, false for
down / off.